// src/types/database.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.18";
  };
  public: {
// ...
    Views: {
      post_feed: {
        Row: {
          content: string | null;
          created_at: string | null;
          dong_code: string | null;
          id: string | null;
          is_hidden: boolean | null;
          is_mine: boolean | null;
          like_count: number | null;
          liked_by_me: boolean | null;
          nickname: string | null;
          tag: string | null;
          user_id: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "posts_dong_code_fkey";
            columns: ["dong_code"];
            isOneToOne: false;
            referencedRelation: "dongs";
            referencedColumns: ["code"];
          },
          {
            foreignKeyName: "posts_user_id_fkey";
            columns: ["user_id"];
            isOneToOne: false;
            referencedRelation: "profiles";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "posts_user_id_fkey";
            columns: ["user_id"];
            isOneToOne: false;
            referencedRelation: "public_profiles";
            referencedColumns: ["id"];
          },
        ];
      };
// ...
    };
// ...
  };
};